+++ src/main.js
@@ -1,3 +1,3 @@
-import { run } from './action-update-license-year';
+const { run } = require('./action-update-license-year');
 
 run();